(MAINT) Update i18n tests for `panic_t!` and `assert_t!`
This change updates the i18n Pester tests to account for:

- Usage of the `t!` macro when the translation key is defined on the next line.
- Usage of the convenience `panic_t!` macro, which simplifies sending translated
  strings to the `panic!` macro.
- Usage of the convenience `assert_t!` macro, which simplifies sending translated
  strings to the panic when an `assert!` macro fails.

These use cases were required for the `dsc-lib-jsonschema` crate, which uses both
`panic_t!` and `assert_t!` to raise translated panic messages when a transform is
invalidly applied.

While this change doesn't modify the behavior for any other crates, it does lay the
groundwork for translating panic messages and assertions throughout the crates.