(MAINT) Disable docs tests for library crates by default
Prior to this change, invoking `cargo test` for library crates without
specifically choosing a test subset (unit, integration, etc) would
always run the documentation tests.

Documentation tests are particularly slow because Rust compiles a binary
for every codeblock in your documentation. This is the slowest part of
the rust tests when working on a library with extensive examples.

This change adds the `doctest` setting to the crate manifests as `false`
to disable documentation tests by default. These tests can still be
invoked by passing the `--doc` flag to `cargo test`.

Future changes will provide simplified access to generating and testing
the docs from the build script, as well as validating documentation in
CI.