(MAINT) Enable generating/testing docs in CI
This change updates the build helpers module to enable separately
generating documentation for Rust crates and testing the examples in
that documentation.

It also adds a new workflow job for generating and testing documentation
separately from the rest of the build and test scripts. While these jobs
failing won't prevent the continued unit/integration/acceptance tests,
it will still fail the build.

This should help ensure that any examples in our documentation remain
valid. It _doesn't_ test to ensure that public APIs are documented, only
that the documentation that exists is minimally valid.