Implement deserializer for `resources` to handle `Array` and `Object`
Which is necessary to support ARMv2's new "symbolic names" structure.

Many alternatives were tried, including updating the type to be an `enum`
and then either implementing the used `Vec`-like interfaces
or pattern matching on the type everywhere `resources` used.
Neither of these were great and resulted in a lot of unnecessary code churn.

The `serde_with` library was also briefly tried with the `PickAs` trait,
but the only thing that gained was eliminating the need to define the
(very easy) deserialization of `Array`,
at the cost of multiple additional libraries.

This seems to be the best way forward as it results in the least possible
potential problems and eliminates changes throughout the codebase.
What it lacks it the ability serialize to ARMv2, but that's not (yet) desired.
It also has not fixed `dependsOn` but that would need to be done regardless.