(GH-538) Add `get_keyword_as_subschema` extension methods
Prior to this change, the `dsc-lib-jsonschema` crate defined
extension methods for schemas to retrieve keywords as various
underlying data types when the caller knows the data type they
need. However, the extension methods didn't include a way to
retrieve values as _schemas_, so any use of those values
requires the caller to reimplement the extension methods logic
or to manually convert the value to a schema.

This change adds two helper methods to retrieve a keyword as
a `schemars::Schema` instance (borrowed and mutably borrowed)
to make working with subschemas more ergonomic.