(MAINT) Reimplement Rust i18n acceptance tests
Prior to this change, the i18n tests used simple string matching to
search for translation keys in the `en-us.toml` file and the project
Rust code.

This change defines a new `DscProjectRustTranslationInfo` class and
`Get-TranslationKey` function in the `dsc_i18n` acceptance tests.

This change reimplements the i18n tests to use the new class and
function for retrieving and analyzing translation strings for Rust
projects, building in support for both TOML and YAML translation files.

This is required to support the translation strings for JSON Schema
annotations, which are lengthy. It enables us to keep the translations
for messages in the code in the `en-us.toml` file and all schema-related
translation strings in separate YAML files, where the hierarchy and
multi-line strings fit more naturally.

This change ensures that we can continue to accurately test our
translation strings, checking:

- That each project with a `locales` folder defines and uses at least
  one translation string.
- That no translation keys are defined more than once across all
  translation files.
- That every defined translation key is used in the Rust code.
- That every translation key the Rust code references is defined in a
  translation file for `en-us`.