(GH-538) Move `DscRepoSchema` into `dsc-lib-jsonschema`
Prior to this change, types in `dsc-lib` depended on the`DscRepoSchema`
trait defined in the same crate for generating the set of recognized
schema URIs and validating them.

This change migrates this JSON Schema specific code into the
`dsc-lib-jsonschema` crate for better organization and reuse, since
this is actually required for any schema published from the DSC
repository, even if those crates don't depend on `dsc-lib`.

This migration also takes advantage of some improvements in code
organization for compilation and testing, migrating the separate
related types, like `SchemaForm` and `SchemaUriPrefix`, into their
own private modules with the parent module re-exporting those types.

Finally, this migration also incorporates a build script, which
checks the git tags for the project to generate the variants for the
`RecognizedSchemaVersion` enum. Prior to this implementation, the
developers were required to manually update the enum definition prior
to every release.

This change does _not_ update any other crates to use the new
implementation. A separate, future change will modify the other crates.