(GH-538) Re-export `dsc-lib-jsonschema` in `dsc-lib`
Prior to this change, the `dsc-lib` crate used `dsc-lib-jsonschema` as
a dependency. With the transition of the `DscRepoSchema` trait and the
related types to the `dsc-lib-jsonschema` crate, it makes more sense to
re-export the dependency from within the `dsc-lib` crate.

This change re-exports the `dsc-lib-jsonschema` crate as
`dsc_lib::schemas`, replacing the prior module and direct references to
`dsc-lib-jsonschema` for transformers and other crate functionality.

Now other crates can choose whether to depend on `dsc-lib` with all of
the associated functionality, or `dsc-lib-jsonschema` if they only need
the JSON Schema related functionality.