(GH-538) Use derive macro for `DscRepoSchema` trait
Prior to this change, the resource manifest, extension manifest, and
configuration document structs in `dsc-lib` manually implemented the
`DscRepoSchema` trait.

This change uses the new macro to derive the trait on those types and
other types in the library that currently map to schema files. This
change also makes an effort to correctly annotate types that aren't
covered by the separate schemas, but as not all types map to a schema
file, these changes don't fully cover all types that derive the
`JsonSchema` trait.

This change prepares the library for further canonicalization steps to
steer the generated schemas towards the definitions in the canonical
schemas.