(MAINT) Pin `tree-sitter-cli` to `0.25.10`
Prior to this change, CI began  failing due to errors in the generated
bindings for the grammar crates. This seems to be related to a new
release of tree-sitter.

This change pins the version of tree-sitter in the build script to use
`0.25.10`, which is the last version that doesn't cause build failures.

We should investigate further in the future, but for now the pinning
should resolve CI failures that are blocking all other PRs.