(MAINT) Update schema for project data files
This change updates the JSON Schemas for the project data files
(`.project.data.json` and `build.data.json`) to recognize the new
`ClippyPedanticUnclean` option.

This change also updates the `build.data.json` file with the latest
project data. This doesn't affect the build, which always checks the
individual project data files first, but does make it easier to check
the project options as a maintainer.